+++ src/ots/OTServer.cpp
@@ -1352,10 +1352,10 @@ void OTServer::Init(bool bReadOnly/*=false*/)
                 // There was a real PID in there.
                 if (old_pid != 0)
                 {
-                    const uint64_t lPID = static_cast<uint64_t>(old_pid);
+                    const uint64_t lPID = old_pid;
                     OTLog::vError("\n\n\nIS OPEN-TRANSACTIONS ALREADY RUNNING?\n\n"
-                                  "I found a PID (%lu) in the data lock file, located at: %s\n\n"
-                                  "If the OT process with PID %lu is truly not running anymore, "
+                                  "I found a PID (%llu) in the data lock file, located at: %s\n\n"
+                                  "If the OT process with PID %llu is truly not running anymore, "
                                   "then just ERASE THAT FILE and then RESTART.\n", lPID, strPIDPath.Get(), lPID);
                     exit(-1);
                 }
@@ -1366,12 +1366,12 @@ void OTServer::Init(bool bReadOnly/*=false*/)
 
             // 3. GET THE CURRENT (ACTUAL) PROCESS ID.
             //
-            uint32_t the_pid = 0;
+            uint64_t the_pid = 0;
 
 #ifdef _WIN32
-            the_pid = static_cast<uint32_t>(GetCurrentProcessId());
+            the_pid = GetCurrentProcessId();
 #else
-            the_pid = static_cast<uint32_t>(getpid());
+            the_pid = getpid();
 #endif
 
             // 4. OPEN THE FILE IN WRITE MODE, AND SAVE THE PID TO IT.
@@ -1384,7 +1384,7 @@ void OTServer::Init(bool bReadOnly/*=false*/)
                 pid_outfile.close();
             }
             else
-                OTLog::vError("Failed trying to open data locking file (to store PID %lu): %s\n",
+                OTLog::vError("Failed trying to open data locking file (to store PID %llu): %s\n",
                               the_pid, strPIDPath.Get());
         }
     }
